fix(endorse): merged Certified+Bluesky search in the Endorse-people modal
The Endorse-people modal searched Bluesky's AppView only, so Certified-
only orgs (e.g. biofi-project) never appeared — the same gap fixed for
the list account-picker. Extract that merged search (Certified indexer +
Bluesky, deduped by DID, handles backfilled) into a shared
`searchMergedActors`, and use it in both surfaces:

- new src/lib/atproto/actor-search.ts (the shared helper)
- profile-lists: its inline merge logic is replaced by the helper
- endorse-people-modal: searches via the helper instead of /api/search-actors

Endorsing only needs the DID, so bsky-only accounts work too — no
profile-record CID is required to endorse.
